refactor: separate prompting from the Generator class for reusability
yeoman generator is designed not to use prompting in sub generators,
so the prompting implementation should be separated from the Generator class.
